requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
SQLAlchemy==2.0.21
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
crcmod==2.1
google-cloud-storage==2.10.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
ruff==0.1.7
mypy==1.7.1
gunicorn==21.2.0
requests==2.31.0
